pyproject.toml
[project]
name = "downloader-tg"
version = "1.0.0"
description = "Telegram bot for downloading videos from various platforms"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.15.0",
    "yt-dlp>=2024.12.13",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
]

[project.scripts]
downloader-bot = "src.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

